gator-daemon: bump to 8.4.0
Gator daemon moved to CMake build system so update the recipe to inherit
cmake and drop patches that has been merged.

This new version have removed the internal libmxml and now required to link with it.
As the libmxml is present in meta-oe and we don't want meta-oe to be an explicit
dependency of meta-arm move the gator daemon recipe to a dynamic-layer.

At the same time remove the Apache-2.0 license as it was set for libmxml.

Signed-off-by: Clément Péron <peron.clem@gmail.com>